backend/ImageProcessor/requirements.txt
Flask
google-cloud-storage
google-cloud-pubsub
google-cloud-firestore
Pillow
python-dotenv
flask_cors
firebase-admin